IT WORKSSSSS

+++ src/util/firmware/include/serial_handling.h
@@ -6,6 +6,6 @@
 // Your code goes here
 extern bool newData;
 
-void recvWithStartEndMarkers(char* messageBuffer, const byte bufferSize);
-bool waitForCommand(const char* cmmdString, char* messageBuffer, const byte bufferSize);
+void recvWithStartEndMarkers(char* messageBuffer, const int bufferSize);
+bool waitForCommand(const char* cmmdString, char* messageBuffer, const int bufferSize);
 #endif // SERIAL_HANDLING_H
